Add transaction_id and logdata_message mapping
The transaction_id and logdata_message keys are
Mapped to the HSDP LogEvent fields.

Important: when the logdata_message key is present
It will override the default behavior of dumping
All the key/values to the logData.Message field.
If you still want to preserve these keys you should
enable `CustomField` option which will store these
Fields in the custom field of the EventLog document